requirements.txt
flask-bcrypt
flask-restx
pyjwt
Flask-Script
flask-cors
python-dotenv
mongoengine